data/requirements.txt
python-dotenv==1.0.1
confluent_kafka==2.3.0
schedule==1.2.1
requests==2.31.0